notebook clean-up
scraping takes about 2h (in batches of 50 pages in parallel)
data extraction took ~1.5h

Scraping pages one by one would take 8-20h. In batches of 50 it is not much faster than in batches of 10. If re-download was ever needed I'd suggest to go with a batch size of 10.

There is no requirements.txt, but the code runs out-of-the-box in google-colab as of April 2021.